+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [unreleased]
+
+### Changed
+
+- Check parameters at initialisation  
+
 ## [5.0.1] 2024-07-12
 
 ### Fixed